Feat: setCurrentPage ->  handleClickNextButton로 Naming 변경
