+++ src/components/survey/Age/Age.tsx
@@ -3,10 +3,10 @@ import React from 'react';
 interface Props {
   nickname: string;
   setAge: (value: number) => void;
-  setCurrentPage: (page: number) => void;
+  handleClickNextButton: (currentPage: number) => void;
 }
 
-const Age = ({ nickname, setAge, setCurrentPage }: Props): JSX.Element => {
+const Age = ({ nickname, setAge, handleClickNextButton }: Props): JSX.Element => {
   return <div />;
 };
