+++ src/components/survey/Nickname/Nickname.tsx
@@ -3,10 +3,10 @@ import React from 'react';
 interface Props {
   nickname: string;
   setNickname: (value: string) => void;
-  setCurrentPage: (page: number) => void;
+  handleClickNextButton: (currentPage: number) => void;
 }
 
-const Nickname = ({ nickname, setNickname, setCurrentPage }: Props): JSX.Element => {
+const Nickname = ({ nickname, setNickname, handleClickNextButton }: Props): JSX.Element => {
   return <div />;
 };
