+++ src/components/survey/Weight/Weight.tsx
@@ -1,0 +1,13 @@
+import React, { useState } from 'react';
+
+interface Props {
+  nickname: string;
+  setWeight: (value: number) => void;
+  handleClickNextButton: (currentPage: number) => void;
+}
+
+function Weight({ nickname, setWeight, handleClickNextButton }: Props): JSX.Element {
+  return <div />;
+}
+
+export default Weight;
