+++ src/components/survey/index.tsx
@@ -1,5 +1,6 @@
 export { default as Age } from './Age/Age';
-export { default as BodyInfo } from './BodyInfo/BodyInfo';
 export { default as Nickname } from './Nickname/Nickname';
 export { default as Gender } from './Gender/Gender';
 export { default as Split } from './Split/Split';
+export { default as Height } from './Height/Height';
+export { default as Weight } from './Weight/Weight';
